refactor: remove `setDefaultOptions`, using `createScopedRequest` instead.
