--- src/options/defaultOptions.ts
@@ -1,6 +1,0 @@
-import { RequestOptions } from '../interfaces';
-
-/**
- * Globals Options.
- */
-export const defaultOptions: RequestOptions = {};
