--- src/options/index.ts
@@ -1,3 +1,0 @@
-export * from './defaultOptions';
-export * from './mergeOptions';
-export * from './setDefaultOptions';
