--- src/options/setDefaultOptions.ts
@@ -1,18 +1,0 @@
-import { RequestOptions } from '../interfaces';
-import { defaultOptions } from './defaultOptions';
-import { mergeOptions } from './mergeOptions';
-
-/**
- * Set Default Fatcher Request Options.
- *
- * Tips:
- *
- * If change default options, will `effect` any fatcher instances.
- *
- * If you want to have a scope options with fatcher. Use `createScopedRequest` instead.
- *
- * Inline Options `>` Scoped Options `>` Default Options
- */
-export function setDefaultOptions<T extends RequestOptions>(patchRequestOptions: Partial<T>) {
-    Object.assign(defaultOptions, mergeOptions(defaultOptions, patchRequestOptions));
-}
